Major refactor
Major refactor of codebase. Seperated to smaller understandable chunks.

Functionality should be exactly the same.

More refactor needed, see def scan_directories_and_create_video in video_creator.py.
